# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the SA Payment service.",
        "title": "SA Payment API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8003",
    "basePath": "/",
    "paths": {
...
        "/api/payment/v1/info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve all payment information records for the authenticated patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment-info"
                ],
                "summary": "List payment information",
                "responses": {
                    "200": {
                        "description": "Payment information retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAllPaymentInfosResponseDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve payment information",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an existing payment information record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment-info"
                ],
                "summary": "Update payment information",
                "parameters": [
                    {
                        "description": "Payment information to update",
                        "name": "payment_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePaymentInfoRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment information updated successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePaymentInfoResponseDto"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Payment information not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update payment information",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new payment method for the authenticated patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment-info"
                ],
                "summary": "Create payment information",
                "parameters": [
                    {
                        "description": "Payment information to create",
                        "name": "payment_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePaymentInfoRequestDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Payment information created successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePaymentInfoResponseDto"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create payment information",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete an existing payment information record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment-info"
                ],
                "summary": "Delete payment information",
                "parameters": [
                    {
                        "description": "Payment information ID to delete",
                        "name": "payment_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeletePaymentInfoRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment information deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.DeletePaymentInfoResponseDto"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Payment information not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete payment information",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/payment/v1/info/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a payment information record by its identifier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment-info"
                ],
                "summary": "Get payment information by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment information ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment information retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.GetPaymentInfoByIDResponseDto"
                        }
                    },
                    "400": {
                        "description": "Invalid payment information ID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Payment information not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve payment information",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreatePaymentInfoRequestDto": {
            "type": "object",
            "required": [
                "details",
                "payment_method"
            ],
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "payment_method": {
                    "enum": [
                        "credit_card",
                        "promptpay"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.PaymentMethod"
                        }
                    ]
                }
            }
        },
        "dto.CreatePaymentInfoResponseDto": {
            "type": "object",
            "properties": {
                "payment_info": {
                    "$ref": "#/definitions/dto.PaymentInfoDto"
                }
            }
        },
...
        "dto.DeletePaymentInfoRequestDto": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.DeletePaymentInfoResponseDto": {
            "type": "object",
            "properties": {
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.GetAllPaymentInfosResponseDto": {
            "type": "object",
            "properties": {
                "delivery_infos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PaymentInfoDto"
                    }
                }
            }
        },
...
        "dto.GetPaymentInfoByIDResponseDto": {
            "type": "object",
            "properties": {
                "payment_info": {
                    "$ref": "#/definitions/dto.PaymentInfoDto"
                }
            }
        },
...
        "dto.PaymentInfoDto": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "string"
                },
                "payment_method": {
                    "$ref": "#/definitions/models.PaymentMethod"
                },
                "user_id": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdatePaymentInfoRequestDto": {
            "type": "object",
            "required": [
                "details",
                "id",
                "payment_method"
            ],
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "string"
                },
                "payment_method": {
                    "enum": [
                        "credit_card",
                        "promptpay"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.PaymentMethod"
                        }
                    ]
                }
            }
        },
        "dto.UpdatePaymentInfoResponseDto": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "string"
                },
                "payment_method": {
                    "$ref": "#/definitions/models.PaymentMethod"
                },
                "user_id": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "models.PaymentMethod": {
            "type": "string",
            "enum": [
                "credit_card",
                "promptpay"
            ],
            "x-enum-varnames": [
                "PaymentMethodCreditCard",
                "PaymentMethodPromptPay"
            ]
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}